refactor: rename `ggufParser` directory to `parser`
