+++ src/gguf/parser/types/GgufFileInfoTypes.ts
