+++ src/gguf/parser/types/GgufMetadataTypes.ts
