+++ src/gguf/parser/utils/GgufReadOffset.ts
