+++ src/gguf/parser/utils/convertMetadataKeyValueRecordToNestedObject.ts
